+++ lib/interfaces/ProcessUploadFileParams.ts
@@ -1,5 +1,6 @@
 export interface ProcessUploadFileParams {
 	languageInferer?: (filePath: string) => Promise<string> | string;
+	filenameInferer?: (filePath: string) => Promise<string> | string;
 	pollStatuses?: boolean;
 	pollInitialWaitTime?: number;
 	pollMaximumWaitTime?: number;