+++ lib/interfaces/UploadTranslationParams.ts
@@ -1,9 +1,9 @@
-import type { UploadFileParams } from "@lokalise/node-api";
 import type { CollectFileParams } from "./CollectFileParams.js";
+import type { PartialUploadFileParams } from "./PartialUploadFileParams.js";
 import type { ProcessUploadFileParams } from "./ProcessUploadFileParams.js";
 
 export interface UploadTranslationParams {
-	uploadFileParams?: Partial<UploadFileParams>;
+	uploadFileParams?: PartialUploadFileParams;
 	collectFileParams?: CollectFileParams;
 	processUploadFileParams?: ProcessUploadFileParams;
 }